splits out CLI from process
